+++ src/client/BaseClient.ts
@@ -3,24 +3,17 @@
  * Licensed under the MIT License.
  */
 
-import {
-    ClientConfiguration,
-    buildClientConfiguration,
-} from "../config/ClientConfiguration";
-import { ICacheStorage } from "../cache/ICacheStorage";
+import { ClientConfiguration, buildClientConfiguration } from "../config/ClientConfiguration";
 import { INetworkModule } from "../network/INetworkModule";
 import { ICrypto } from "../crypto/ICrypto";
 import { Authority } from "../authority/Authority";
 import { Logger } from "../logger/Logger";
 import { AADServerParamKeys, Constants, HeaderNames } from "../utils/Constants";
 import { NetworkResponse } from "../network/NetworkManager";
-import { ServerAuthorizationTokenResponse } from "../server/ServerAuthorizationTokenResponse";
-import { B2cAuthority } from "../authority/B2cAuthority";
-import { UnifiedCacheManager } from "../unifiedCache/UnifiedCacheManager";
-import { AccountEntity } from "../unifiedCache/entities/AccountEntity";
-import { IAccount } from "../account/IAccount";
-import { AccountCache } from "../unifiedCache/utils/CacheTypes";
-import { CacheHelper } from "../unifiedCache/utils/CacheHelper";
+import { ServerAuthorizationTokenResponse } from "../response/ServerAuthorizationTokenResponse";
+import { TrustedAuthority } from "../authority/TrustedAuthority";
+import { CacheManager } from "../cache/CacheManager";
+import { ServerTelemetryManager } from "../telemetry/server/ServerTelemetryManager";
 
 /**
  * Base application class which will construct requests to send to and handle responses from the Microsoft STS using the authorization code flow.
@@ -36,19 +29,16 @@ export abstract class BaseClient {
     protected cryptoUtils: ICrypto;
 
     // Storage Interface
-    protected cacheStorage: ICacheStorage;
+    protected cacheManager: CacheManager;
 
     // Network Interface
     protected networkClient: INetworkModule;
 
-    // Helper API object for serialized cache operations
-    protected unifiedCacheManager: UnifiedCacheManager;
-
-    // Account object
-    protected account: AccountEntity;
+    // Server Telemetry Manager
+    protected serverTelemetryManager: ServerTelemetryManager;
 
     // Default authority object
-    protected defaultAuthority: Authority;
+    protected authority: Authority;
 
     protected constructor(configuration: ClientConfiguration) {
         // Set the configuration
@@ -61,30 +51,30 @@ export abstract class BaseClient {
         this.cryptoUtils = this.config.cryptoInterface;
 
         // Initialize storage interface
-        this.cacheStorage = this.config.storageInterface;
-
-        // Initialize serialized cache manager
-        this.unifiedCacheManager = new UnifiedCacheManager(
-            this.cacheStorage,
-            this.config.systemOptions.storeInMemory
-        );
+        this.cacheManager = this.config.storageInterface;
 
         // Set the network interface
         this.networkClient = this.config.networkInterface;
 
-        B2cAuthority.setKnownAuthorities(
-            this.config.authOptions.knownAuthorities
-        );
+        // Set TelemetryManager
+        this.serverTelemetryManager = this.config.serverTelemetryManager;
 
-        this.defaultAuthority = this.config.authOptions.authority;
+        TrustedAuthority.setTrustedAuthoritiesFromConfig(this.config.authOptions.knownAuthorities, this.config.authOptions.cloudDiscoveryMetadata);
+
+        this.authority = this.config.authOptions.authority;
     }
 
     /**
      * Creates default headers for requests to token endpoint
      */
-    protected createDefaultTokenRequestHeaders(): Map<string, string> {
+    protected createDefaultTokenRequestHeaders(): Record<string, string> {
         const headers = this.createDefaultLibraryHeaders();
-        headers.set(HeaderNames.CONTENT_TYPE, Constants.URL_FORM_CONTENT_TYPE);
+        headers[HeaderNames.CONTENT_TYPE] = Constants.URL_FORM_CONTENT_TYPE;
+
+        if (this.serverTelemetryManager) {
+            headers[HeaderNames.X_CLIENT_CURR_TELEM] = this.serverTelemetryManager.generateCurrentRequestHeaderValue();
+            headers[HeaderNames.X_CLIENT_LAST_TELEM] = this.serverTelemetryManager.generateLastRequestHeaderValue();
+        }
 
         return headers;
     }
@@ -92,14 +82,14 @@ export abstract class BaseClient {
     /**
      * addLibraryData
      */
-    protected createDefaultLibraryHeaders(): Map<string, string> {
-        const headers = new Map<string, string>();
+    protected createDefaultLibraryHeaders(): Record<string, string> {
+        const headers: Record<string, string> = {};
 
         // client info headers
-        headers.set(`${AADServerParamKeys.X_CLIENT_SKU}`,this.config.libraryInfo.sku);
-        headers.set(`${AADServerParamKeys.X_CLIENT_VER}`, this.config.libraryInfo.version);
-        headers.set(`${AADServerParamKeys.X_CLIENT_OS}`, this.config.libraryInfo.os);
-        headers.set(`${AADServerParamKeys.X_CLIENT_CPU}`, this.config.libraryInfo.cpu);
+        headers[AADServerParamKeys.X_CLIENT_SKU] = this.config.libraryInfo.sku;
+        headers[AADServerParamKeys.X_CLIENT_VER] = this.config.libraryInfo.version;
+        headers[AADServerParamKeys.X_CLIENT_OS] = this.config.libraryInfo.os;
+        headers[AADServerParamKeys.X_CLIENT_CPU] = this.config.libraryInfo.cpu;
 
         return headers;
     }
@@ -110,30 +100,19 @@ export abstract class BaseClient {
      * @param queryString
      * @param headers
      */
-    protected executePostToTokenEndpoint(tokenEndpoint: string, queryString: string, headers: Map<string, string>): Promise<NetworkResponse<ServerAuthorizationTokenResponse>> {
-        return this.networkClient.sendPostRequestAsync<
+    protected async executePostToTokenEndpoint(tokenEndpoint: string, queryString: string, headers: Record<string, string>): Promise<NetworkResponse<ServerAuthorizationTokenResponse>> {
+        const response = await this.networkClient.sendPostRequestAsync<
         ServerAuthorizationTokenResponse
         >(tokenEndpoint, {
             body: queryString,
             headers: headers,
         });
-    }
 
-    /**
-     * Get all currently signed in accounts.
-     */
-    public getAllAccounts(): IAccount[] {
-        const currentAccounts: AccountCache = this.unifiedCacheManager.getAllAccounts();
-        const accountValues: AccountEntity[] = Object.values(currentAccounts);
-        const numAccounts = accountValues.length;
-        if (numAccounts < 1) {
-            return null;
-        } else {
-            const allAccounts = accountValues.map<IAccount>((value) => {
-                const accountObj: AccountEntity = JSON.parse(JSON.stringify(value));
-                return CacheHelper.toIAccount(accountObj);
-            });
-            return allAccounts;
+        if (this.config.serverTelemetryManager && response.status < 500 && response.status !== 429) {
+            // Telemetry data successfully logged by server, clear Telemetry cache
+            this.config.serverTelemetryManager.clearTelemetryCache();
         }
+
+        return response;
     }
 }