+++ src/client/RefreshTokenClient.ts
@@ -7,8 +7,8 @@ import { ClientConfiguration } from "../config/ClientConfiguration";
 import { BaseClient } from "./BaseClient";
 import { RefreshTokenRequest } from "../request/RefreshTokenRequest";
 import { Authority, NetworkResponse } from "..";
-import { ServerAuthorizationTokenResponse } from "../server/ServerAuthorizationTokenResponse";
-import { RequestParameterBuilder } from "../server/RequestParameterBuilder";
+import { ServerAuthorizationTokenResponse } from "../response/ServerAuthorizationTokenResponse";
+import { RequestParameterBuilder } from "../request/RequestParameterBuilder";
 import { ScopeSet } from "../request/ScopeSet";
 import { GrantType } from "../utils/Constants";
 import { ResponseHandler } from "../response/ResponseHandler";
@@ -24,19 +24,19 @@ export class RefreshTokenClient extends BaseClient {
     }
 
     public async acquireToken(request: RefreshTokenRequest): Promise<AuthenticationResult>{
-        const response = await this.executeTokenRequest(request, this.defaultAuthority);
+        const response = await this.executeTokenRequest(request, this.authority);
 
         const responseHandler = new ResponseHandler(
             this.config.authOptions.clientId,
-            this.unifiedCacheManager,
+            this.cacheManager,
             this.cryptoUtils,
             this.logger
         );
 
         responseHandler.validateTokenResponse(response.body);
-        const tokenResponse = await responseHandler.generateAuthenticationResult(
+        const tokenResponse = responseHandler.handleServerTokenResponse(
             response.body,
-            this.defaultAuthority
+            this.authority
         );
 
         return tokenResponse;
@@ -46,7 +46,7 @@ export class RefreshTokenClient extends BaseClient {
         : Promise<NetworkResponse<ServerAuthorizationTokenResponse>> {
 
         const requestBody = this.createTokenRequestBody(request);
-        const headers: Map<string, string> = this.createDefaultTokenRequestHeaders();
+        const headers: Record<string, string> = this.createDefaultTokenRequestHeaders();
 
         return this.executePostToTokenEndpoint(authority.tokenEndpoint, requestBody, headers);
     }
@@ -54,13 +54,29 @@ export class RefreshTokenClient extends BaseClient {
     private createTokenRequestBody(request: RefreshTokenRequest): string {
         const parameterBuilder = new RequestParameterBuilder();
 
-        const scopeSet = new ScopeSet(request.scopes || [],
-            this.config.authOptions.clientId,
-            false);
-        parameterBuilder.addScopes(scopeSet);
         parameterBuilder.addClientId(this.config.authOptions.clientId);
+
+        const scopeSet = new ScopeSet(request.scopes || []);
+        parameterBuilder.addScopes(scopeSet);
+        
         parameterBuilder.addGrantType(GrantType.REFRESH_TOKEN_GRANT);
+
+        parameterBuilder.addClientInfo();
+
+        const correlationId = request.correlationId || this.config.cryptoInterface.createNewGuid();
+        parameterBuilder.addCorrelationId(correlationId);
+
         parameterBuilder.addRefreshToken(request.refreshToken);
+
+        if (this.config.clientCredentials.clientSecret) {
+            parameterBuilder.addClientSecret(this.config.clientCredentials.clientSecret);
+        }
+
+        if (this.config.clientCredentials.clientAssertion) {
+            const clientAssertion = this.config.clientCredentials.clientAssertion;
+            parameterBuilder.addClientAssertion(clientAssertion.assertion);
+            parameterBuilder.addClientAssertionType(clientAssertion.assertionType);
+        }
 
         return parameterBuilder.createQueryString();
     }