+++ src/client/SilentFlowClient.ts
@@ -8,17 +8,20 @@ import { ClientConfiguration } from "../config/ClientConfiguration";
 import { SilentFlowRequest } from "../request/SilentFlowRequest";
 import { AuthenticationResult } from "../response/AuthenticationResult";
 import { CredentialType } from "../utils/Constants";
-import { IdTokenEntity } from "../unifiedCache/entities/IdTokenEntity";
-import { CacheHelper } from "../unifiedCache/utils/CacheHelper";
-import { AccessTokenEntity } from "../unifiedCache/entities/AccessTokenEntity";
-import { RefreshTokenEntity } from "../unifiedCache/entities/RefreshTokenEntity";
+import { IdTokenEntity } from "../cache/entities/IdTokenEntity";
+import { AccessTokenEntity } from "../cache/entities/AccessTokenEntity";
+import { RefreshTokenEntity } from "../cache/entities/RefreshTokenEntity";
 import { ScopeSet } from "../request/ScopeSet";
 import { IdToken } from "../account/IdToken";
 import { TimeUtils } from "../utils/TimeUtils";
 import { RefreshTokenRequest } from "../request/RefreshTokenRequest";
 import { RefreshTokenClient } from "./RefreshTokenClient";
 import { ClientAuthError } from "../error/ClientAuthError";
-import { CredentialFilter, CredentialCache } from "../unifiedCache/utils/CacheTypes";
+import { CredentialFilter, CredentialCache } from "../cache/utils/CacheTypes";
+import { AccountEntity } from "../cache/entities/AccountEntity";
+import { CredentialEntity } from "../cache/entities/CredentialEntity";
+import { ClientConfigurationError } from "../error/ClientConfigurationError";
+import { ResponseHandler } from "../response/ResponseHandler";
 
 export class SilentFlowClient extends BaseClient {
 
@@ -32,22 +35,26 @@ export class SilentFlowClient extends BaseClient {
      * @param request
      */
     public async acquireToken(request: SilentFlowRequest): Promise<AuthenticationResult> {
+        // Cannot renew token if no request object is given.
+        if (!request) {
+            throw ClientConfigurationError.createEmptyTokenRequestError();
+        }
+        
         // We currently do not support silent flow for account === null use cases; This will be revisited for confidential flow usecases
         if (!request.account) {
             throw ClientAuthError.createNoAccountInSilentRequestError();
         } 
 
-        const requestScopes = new ScopeSet(request.scopes || [], this.config.authOptions.clientId, true);
-        // fetch account
-        const accountKey: string = CacheHelper.generateAccountCacheKey(request.account);
-        const cachedAccount = this.unifiedCacheManager.getAccount(accountKey);
+        const requestScopes = new ScopeSet(request.scopes || []);
+    
+        // Get account object for this request.
+        const accountKey: string = AccountEntity.generateAccountCacheKey(request.account);
+        const cachedAccount = this.cacheManager.getAccount(accountKey);
 
         const homeAccountId = cachedAccount.homeAccountId;
         const environment = cachedAccount.environment;
 
-        // fetch idToken, accessToken, refreshToken
-        const cachedIdToken = this.readIdTokenFromCache(homeAccountId, environment, cachedAccount.realm);
-        const idTokenObj = new IdToken(cachedIdToken.secret, this.config.cryptoInterface);
+        // Get current cached tokens
         const cachedAccessToken = this.readAccessTokenFromCache(homeAccountId, environment, requestScopes, cachedAccount.realm);
         const cachedRefreshToken = this.readRefreshTokenFromCache(homeAccountId, environment);
 
@@ -60,27 +67,24 @@ export class SilentFlowClient extends BaseClient {
 
             const refreshTokenClient = new RefreshTokenClient(this.config);
             const refreshTokenRequest: RefreshTokenRequest = {
-                scopes: request.scopes,
-                refreshToken: cachedRefreshToken.secret,
-                authority: request.authority
+                ...request,
+                refreshToken: cachedRefreshToken.secret
             };
 
             return refreshTokenClient.acquireToken(refreshTokenRequest);
         }
 
-        // generate Authentication Result
-        return {
-            uniqueId: idTokenObj.claims.oid || idTokenObj.claims.sub,
-            tenantId: idTokenObj.claims.tid,
-            scopes: requestScopes.asArray(),
-            account: CacheHelper.toIAccount(cachedAccount),
-            idToken: cachedIdToken.secret,
-            idTokenClaims: idTokenObj.claims,
-            accessToken: cachedAccessToken.secret,
-            expiresOn: new Date(cachedAccessToken.expiresOn),
-            extExpiresOn: new Date(cachedAccessToken.extendedExpiresOn),
-            familyId: null,
-        };
+        // Return tokens from cache
+        this.config.serverTelemetryManager.incrementCacheHits();
+        const cachedIdToken = this.readIdTokenFromCache(homeAccountId, environment, cachedAccount.realm);
+        const idTokenObj = new IdToken(cachedIdToken.secret, this.config.cryptoInterface);
+
+        return ResponseHandler.generateAuthenticationResult({
+            account: cachedAccount,
+            accessToken: cachedAccessToken,
+            idToken: cachedIdToken,
+            refreshToken: cachedRefreshToken
+        }, idTokenObj, true);
     }
 
     /**
@@ -88,14 +92,14 @@ export class SilentFlowClient extends BaseClient {
      * @param request
      */
     private readIdTokenFromCache(homeAccountId: string, environment: string, inputRealm: string): IdTokenEntity {
-        const idTokenKey: string = CacheHelper.generateCredentialCacheKey(
+        const idTokenKey: string = CredentialEntity.generateCredentialCacheKey(
             homeAccountId,
             environment,
             CredentialType.ID_TOKEN,
             this.config.authOptions.clientId,
             inputRealm
         );
-        return this.unifiedCacheManager.getCredential(idTokenKey) as IdTokenEntity;
+        return this.cacheManager.getCredential(idTokenKey) as IdTokenEntity;
     }
 
     /**
@@ -110,10 +114,10 @@ export class SilentFlowClient extends BaseClient {
             credentialType: CredentialType.ACCESS_TOKEN,
             clientId: this.config.authOptions.clientId,
             realm: inputRealm,
-            target: scopes.printScopes()
+            target: scopes.printScopesLowerCase()
         };
-        const credentialCache: CredentialCache = this.unifiedCacheManager.getCredentialsFilteredBy(accessTokenFilter);
-        const accessTokens = Object.values(credentialCache.accessTokens);
+        const credentialCache: CredentialCache = this.cacheManager.getCredentialsFilteredBy(accessTokenFilter);
+        const accessTokens = Object.keys(credentialCache.accessTokens).map(key => credentialCache.accessTokens[key]);
         if (accessTokens.length > 1) {
             // TODO: Figure out what to throw or return here.
         } else if (accessTokens.length < 1) {
@@ -127,13 +131,13 @@ export class SilentFlowClient extends BaseClient {
      * @param request
      */
     private readRefreshTokenFromCache(homeAccountId: string, environment: string): RefreshTokenEntity {
-        const refreshTokenKey: string = CacheHelper.generateCredentialCacheKey(
+        const refreshTokenKey: string = CredentialEntity.generateCredentialCacheKey(
             homeAccountId,
             environment,
             CredentialType.REFRESH_TOKEN,
             this.config.authOptions.clientId
         );
-        return this.unifiedCacheManager.getCredential(refreshTokenKey) as RefreshTokenEntity;
+        return this.cacheManager.getCredential(refreshTokenKey) as RefreshTokenEntity;
     }
 
     /**