+++ src/url/UrlString.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-import { ServerAuthorizationCodeResponse } from "../server/ServerAuthorizationCodeResponse";
+import { ServerAuthorizationCodeResponse } from "../response/ServerAuthorizationCodeResponse";
 import { ClientConfigurationError } from "../error/ClientConfigurationError";
 import { ClientAuthError } from "../error/ClientAuthError";
 import { StringUtils } from "../utils/StringUtils";
@@ -22,11 +22,13 @@ export class UrlString {
     
     constructor(url: string) {
         this._urlString = url;
-        if (!StringUtils.isEmpty(this._urlString) && StringUtils.isEmpty(this.getHash())) {
-            this._urlString = this.canonicalizeUri(url);
-        } else if (StringUtils.isEmpty(this._urlString)) {
+        if (StringUtils.isEmpty(this._urlString)) {
             // Throws error if url is empty
             throw ClientConfigurationError.createUrlEmptyError();
+        }
+
+        if (StringUtils.isEmpty(this.getHash())) {
+            this._urlString = UrlString.canonicalizeUri(url);
         }
     }
 
@@ -34,12 +36,12 @@ export class UrlString {
      * Ensure urls are lower case and end with a / character.
      * @param url 
      */
-    private canonicalizeUri(url: string): string {
+    static canonicalizeUri(url: string): string {
         if (url) {
             url = url.toLowerCase();
         }
 
-        if (url && !url.endsWith("/")) {
+        if (url && !StringUtils.endsWith(url, "/")) {
             url += "/";
         }
 
@@ -86,6 +88,10 @@ export class UrlString {
         return this.urlString;
     }
 
+    static removeHashFromUrl(url: string): string {
+        return UrlString.canonicalizeUri(url.split("#")[0]);
+    }
+
     /**
      * Given a url like https://a:b/common/d?e=f#g, and a tenantId, returns https://a:b/tenantId/d
      * @param href The url
@@ -104,26 +110,7 @@ export class UrlString {
      * Returns the anchor part(#) of the URL
      */
     getHash(): string {
-        const hashIndex1 = this.urlString.indexOf("#");
-        const hashIndex2 = this.urlString.indexOf("#/");
-        if (hashIndex2 > -1) {
-            return this.urlString.substring(hashIndex2 + 2);
-        } else if (hashIndex1 > -1) {
-            return this.urlString.substring(hashIndex1 + 1);
-        }
-        return "";
-    }
-
-    /**
-     * Returns deserialized portion of URL hash
-     */
-    getDeserializedHash<T>(): T {
-        const hash = this.getHash();
-        const deserializedHash: T = StringUtils.queryStringToObject<T>(hash);
-        if (!deserializedHash) {
-            throw ClientAuthError.createHashNotDeserializedError(JSON.stringify(deserializedHash));
-        }
-        return deserializedHash;
+        return UrlString.parseHash(this.urlString);
     }
 
     /**
@@ -153,20 +140,67 @@ export class UrlString {
         return urlComponents;
     }
 
+    static getDomainFromUrl(url: string): string {
+        const regEx = RegExp("^([^:/?#]+://)?([^/?#]*)");
+
+        const match = url.match(regEx);
+
+        if (!match) {
+            throw ClientConfigurationError.createUrlParseError(`Given url string: ${url}`);
+        }
+
+        return match[2];
+    }
+    
+    /**
+     * Parses hash string from given string. Returns empty string if no hash symbol is found.
+     * @param hashString 
+     */
+    static parseHash(hashString: string): string {
+        const hashIndex1 = hashString.indexOf("#");
+        const hashIndex2 = hashString.indexOf("#/");
+        if (hashIndex2 > -1) {
+            return hashString.substring(hashIndex2 + 2);
+        } else if (hashIndex1 > -1) {
+            return hashString.substring(hashIndex1 + 1);
+        }
+        return "";
+    }
+
     static constructAuthorityUriFromObject(urlObject: IUri): UrlString {
         return new UrlString(urlObject.Protocol + "//" + urlObject.HostNameAndPort + "/" + urlObject.PathSegments.join("/"));
     }
 
     /**
+     * Returns URL hash as server auth code response object.
+     */
+    static getDeserializedHash(hash: string): ServerAuthorizationCodeResponse {
+        // Check if given hash is empty
+        if (StringUtils.isEmpty(hash)) {
+            return {};
+        }
+        // Strip the # symbol if present
+        const parsedHash = UrlString.parseHash(hash);
+        // If # symbol was not present, above will return empty string, so give original hash value
+        const deserializedHash: ServerAuthorizationCodeResponse = StringUtils.queryStringToObject<ServerAuthorizationCodeResponse>(StringUtils.isEmpty(parsedHash) ? hash : parsedHash);
+        // Check if deserialization didn't work
+        if (!deserializedHash) {
+            throw ClientAuthError.createHashNotDeserializedError(JSON.stringify(deserializedHash));
+        }
+        return deserializedHash;
+    }
+
+    /**
      * Check if the hash of the URL string contains known properties
      */
-    static hashContainsKnownProperties(url: string): boolean {
-        if (StringUtils.isEmpty(url)) {
+    static hashContainsKnownProperties(hash: string): boolean {
+        if (StringUtils.isEmpty(hash)) {
             return false;
         }
-        const urlString = new UrlString(url);
-        const parameters = urlString.getDeserializedHash<ServerAuthorizationCodeResponse>();
+
+        const parameters: ServerAuthorizationCodeResponse = UrlString.getDeserializedHash(hash);
         return !!(
+            parameters.code ||
             parameters.error_description ||
             parameters.error ||
             parameters.state