+++ test/error/ClientAuthError.spec.ts
@@ -163,7 +163,7 @@ describe("ClientAuthError.ts Class Unit Tests", () => {
     });
 
     it("createNoTokensFoundError creates a ClientAuthError object", () => {
-        const err: ClientAuthError = ClientAuthError.createNoTokensFoundError("scope1 scope2");
+        const err: ClientAuthError = ClientAuthError.createNoTokensFoundError();
 
         expect(err instanceof ClientAuthError).to.be.true;
         expect(err instanceof AuthError).to.be.true;
@@ -301,6 +301,45 @@ describe("ClientAuthError.ts Class Unit Tests", () => {
         expect(err.errorCode).to.equal(ClientAuthErrorMessage.DeviceCodeExpired.code);
         expect(err.errorMessage).to.include(ClientAuthErrorMessage.DeviceCodeExpired.desc);
         expect(err.message).to.include(ClientAuthErrorMessage.DeviceCodeExpired.desc);
+        expect(err.name).to.equal("ClientAuthError");
+        expect(err.stack).to.include("ClientAuthError.spec.ts");
+    });
+
+    it("createInvalidCacheTypeError creates a ClientAuthError object", () => {
+        const err: ClientAuthError = ClientAuthError.createInvalidCacheTypeError();
+
+        expect(err instanceof ClientAuthError).to.be.true;
+        expect(err instanceof AuthError).to.be.true;
+        expect(err instanceof Error).to.be.true;
+        expect(err.errorCode).to.equal(ClientAuthErrorMessage.invalidCacheType.code);
+        expect(err.errorMessage).to.include(ClientAuthErrorMessage.invalidCacheType.desc);
+        expect(err.message).to.include(ClientAuthErrorMessage.invalidCacheType.desc);
+        expect(err.name).to.equal("ClientAuthError");
+        expect(err.stack).to.include("ClientAuthError.spec.ts");
+    });
+
+    it("createUnexpectedAccountTypeError creates a ClientAuthError object", () => {
+        const err: ClientAuthError = ClientAuthError.createUnexpectedAccountTypeError();
+
+        expect(err instanceof ClientAuthError).to.be.true;
+        expect(err instanceof AuthError).to.be.true;
+        expect(err instanceof Error).to.be.true;
+        expect(err.errorCode).to.equal(ClientAuthErrorMessage.unexpectedAccountType.code);
+        expect(err.errorMessage).to.include(ClientAuthErrorMessage.unexpectedAccountType.desc);
+        expect(err.message).to.include(ClientAuthErrorMessage.unexpectedAccountType.desc);
+        expect(err.name).to.equal("ClientAuthError");
+        expect(err.stack).to.include("ClientAuthError.spec.ts");
+    });
+
+    it("createUnexpectedCredentialTypeError creates a ClientAuthError object", () => {
+        const err: ClientAuthError = ClientAuthError.createUnexpectedCredentialTypeError();
+
+        expect(err instanceof ClientAuthError).to.be.true;
+        expect(err instanceof AuthError).to.be.true;
+        expect(err instanceof Error).to.be.true;
+        expect(err.errorCode).to.equal(ClientAuthErrorMessage.unexpectedCredentialType.code);
+        expect(err.errorMessage).to.include(ClientAuthErrorMessage.unexpectedCredentialType.desc);
+        expect(err.message).to.include(ClientAuthErrorMessage.unexpectedCredentialType.desc);
         expect(err.name).to.equal("ClientAuthError");
         expect(err.stack).to.include("ClientAuthError.spec.ts");
     });